Refactor config script

+++ config.sh
@@ -3,5 +3,8 @@ ln -sf "$(pwd)/dotfiles/vim/.vimrc" ~/.vimrc
 ln -sf "$(pwd)/dotfiles/tmux/.tmux.conf" ~/.tmux.conf
 ln -sf "$(pwd)/dotfiles/ghostty/config" ~/.config/ghostty/config
 
-rm -r ~/.config/nvim
+if [ -d ~/.config/nvim ]; then
+  rm -r ~/.config/nvim
+fi
+
 cp -rs "$(pwd)/dotfiles/nvim" ~/.config/nvim
